Refactor AddImageService to accept Dio instance as a parameter
The AddImageService class has been refactored to accept a Dio instance as a parameter instead of creating a new instance within the class. This change improves the flexibility and testability of the class. The MedicalTestsView has been updated to pass a Dio instance to the AddImageService constructor. The PredictionView has been updated to remove unnecessary code.