Refactor authentication flow and add user details to state
This commit refactors the authentication flow by adding user details to the state. The `AuthCubit` now stores the user's name, email, and `User` object after successful login or signup. Additionally, the `userSignupWithEmailAndPassword` method now takes a `userName` parameter to store the user's name in the Firestore database.

The `HomeStack` widget now retrieves the user's name from the `AuthCubit` state to display a personalized greeting.

The `UserViewBodyUserDetails` widget also retrieves the user's name and email from the `AuthCubit` state to display in the UI.

Overall, this commit improves the user experience by personalizing the app for each user and simplifies the authentication flow by storing user details in the state.